requirements.txt
# Stock Trading System Dependencies

# Data manipulation and analysis
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0

# Alpaca Trading API
alpaca-py>=0.8.0

# Environment management
python-dotenv>=1.0.0

# Optional: For enhanced analysis
matplotlib>=3.7.0
seaborn>=0.12.0

